requirements.txt
# This list of dependencies was generated with `pip3 freeze`

# Logging and parse args
absl-py==1.0.0

# Gym environment
gym==0.23.1

# PyTorch
torch==1.11.0
torchvision==0.12.0

# Utilities
numpy==1.21.6

# Coding styles and others
pre-commit==2.17.0
black==22.1.0
